Extract hooks from App.jsx
- Extract Aladin integration, sidebar controls, galaxy search, tab management, and survey controls into custom hooks
- Organise hooks into subdirectories (data/, modal/, gallery/, aladin/, ui/)
- Organise components into logical subdirectories (aladin/, pipeline/, ui/, upload/)
- Organise utils into subdirectories (aladin/, constants/, file/, gallery/, processing/, ui/)
• Rename UtilsAladin.jsx → aladinUtils.js and maps/ → aladin/ for consistency
• Optimize imports and remove unused dependencies
